Feat d p

+++ components/asic/asic.c
@@ -58,7 +58,7 @@ int ASIC_set_max_baud(GlobalState * GLOBAL_STATE) {
 }
 
 // .set_difficulty_mask_fn = BM1366_set_job_difficulty_mask,
-void ASIC_set_job_difficulty_mask(GlobalState * GLOBAL_STATE, uint8_t mask) {
+void ASIC_set_job_difficulty_mask(GlobalState * GLOBAL_STATE, uint32_t mask) {
     switch (GLOBAL_STATE->device_model) {
         case BITFORGE_NANO:
             BM1370_set_job_difficulty_mask(mask);
